RetailStore_DataPipeline/Project/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-databricks==7.7.1
apache-airflow-providers-fab

